requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.12.0
aiosignal==1.4.0
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==3.2.2
boto3==1.40.61
botocore==1.40.61
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.120.2
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
xlrd==2.0.2
yarl==1.22.0